+++ lib/celluloid/fsm.rb
@@ -55,6 +55,14 @@ module Celluloid
           name = name.to_sym
           default_state name if options["default"]
           states[name] = State.new(name, options["to"], &block)
+          instance_eval do
+            unless respond_to?("#{name}?")
+              define_method("#{name}?") { current_state.name == name }
+            end
+            unless respond_to?("#{name}!")
+              define_method("#{name}!") { transition name }
+            end
+          end
         end
       end
     end