+++ spec/support/examples/test_machine_class.rb
@@ -23,3 +23,10 @@ class CustomDefaultMachine
 
   default_state :foobar
 end
+
+class AnotherCustomDefaultMachine
+  include Celluloid::FSM
+
+  state :foo, default: true
+  state :bar
+end